backend/pyproject.toml
[project]
name = "javapatching"
version = "0.1.0"
description = "Java Patching Application - JDK version upgrade impact analyzer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "tree-sitter>=0.21.0",
    "tree-sitter-java>=0.21.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.2",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "gitpython>=3.1.41",
    "cryptography>=42.0.0",
    "paramiko>=3.4.0",
    "fastmcp>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
